tests/e2e/requirements.txt
pytest-playwright>=0.6
pyyaml>=6.0
